Move BladePage class into Models\Pages namespace
