+++ packages/framework/src/Services/CollectionService.php
@@ -4,7 +4,7 @@ namespace Hyde\Framework\Services;
 
 use Hyde\Framework\Contracts\AbstractPage;
 use Hyde\Framework\Hyde;
-use Hyde\Framework\Models\BladePage;
+use Hyde\Framework\Models\Pages\BladePage;
 use Hyde\Framework\Models\DocumentationPage;
 use Hyde\Framework\Models\Pages\MarkdownPage;
 use Hyde\Framework\Models\Pages\MarkdownPost;
