+++ packages/framework/src/StaticPageBuilder.php
@@ -5,7 +5,7 @@ namespace Hyde\Framework;
 use Hyde\Framework\Actions\MarkdownConverter;
 use Hyde\Framework\Concerns\InteractsWithDirectories;
 use Hyde\Framework\Contracts\PageContract;
-use Hyde\Framework\Models\BladePage;
+use Hyde\Framework\Models\Pages\BladePage;
 use Hyde\Framework\Models\DocumentationPage;
 use Hyde\Framework\Models\Pages\MarkdownPage;
 use Hyde\Framework\Models\Pages\MarkdownPost;
